updates to spplan, and spplan_* for python subpackage
updates to spplan, spplan_epoch, spplan_target, and spplan_trace for python subpackage. Added support for new fieldgroups and output reorganization. Updates to spplan_target to support using targeting flags rather than the database and changes to use the same coadd name for both observatories with the observatories added to the pseudo field names for the paths.